+++ src/components/footer/index.js
@@ -1,1 +1,3 @@
 import "./footer.css";
+
+export { footerState, footerMethods } from "./footer.js";